address review comments
- gate multiple directional lights behind the new
  Engine::Config::enableMultipleDirectionalLights (default false), so that
  existing scenes containing several directional lights keep using only the
  dominant one; plumbed through the Java and JavaScript bindings
- only use a constant-bounded loop with a guard at feature level 0, where
  GLSL ES 1.00 requires it; bound the loop by the light count otherwise
- store the extra directional light count as a uint8_t
- update the LightManager documentation in the Java bindings

+++ filament/src/details/Scene.h
@@ -195,7 +195,7 @@ public:
                 extraDirectionalLightDirections{};
         std::array<FLightManager::Instance, CONFIG_MAX_EXTRA_DIRECTIONAL_LIGHTS>
                 extraDirectionalLightInstances{};
-        size_t extraDirectionalLightCount = 0;
+        uint8_t extraDirectionalLightCount = 0;
         bool hasContactShadows = false;
     };
